refactor: no longer consider config properties classes as tight to serialization by relocating them
