+++ src/patch-place-break-core/src/main/java/fr/djaytan/mc/jrppb/core/config/properties_v2/ConnectionPoolConfigProperties.java
@@ -20,7 +20,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package fr.djaytan.mc.jrppb.core.config.serialization.properties;
+package fr.djaytan.mc.jrppb.core.config.properties_v2;
 
 import fr.djaytan.mc.jrppb.core.storage.properties.ConnectionPoolProperties;
 import org.jetbrains.annotations.NotNull;
